+++ QCodeEditor/include/internal/QCodeEditor.hpp
@@ -59,6 +59,9 @@ public:
     bool getSeparateMinus() const;
     void setSeparateMinus(bool separateMinus);
 
+    bool highlightBlocks() const;
+    void setHighlightBlocks(bool newHighlightBlocks);
+
 signals:
     void saveTriggered();
     void runEmbeddedTriggered();
@@ -66,6 +69,7 @@ signals:
     void stopTriggered();
     void clearConsoleTriggered();
     void searchTriggered();
+    void runBlockTriggered(QString text);
 
 public Q_SLOTS:
 
@@ -132,6 +136,7 @@ protected:
      * 4. Auto parenthesis
      */
     void keyPressEvent(QKeyEvent* e) override;
+    void keyReleaseEvent(QKeyEvent* e) override;
 
     /**
      * @brief Method, that's called on focus into widget.
@@ -192,7 +197,9 @@ private:
      * @brief Method, that adds highlighting of
      * parenthesis if available.
      */
-    void highlightParenthesis(QList<QTextEdit::ExtraSelection>& extraSelection);
+    void highlightParenthesis(
+            QList<QTextEdit::ExtraSelection>& extraSelection,
+            bool selectBlock);
 
     void highlightSearch(QList<QTextEdit::ExtraSelection>& extraSelection);
 
@@ -222,6 +229,8 @@ private:
     QString m_indentStartStr;
     QString m_indentEndStr;
     bool m_separateMinus;
+    bool m_highlightBlocks;
+    QString m_highlightedBlock;
 
 };