+++ res/CHANGELOG.md
@@ -8,6 +8,8 @@
 * Sanity-check motor parameters on write.
 * Log analysis graph and table alignment fix.
 * Disable BMS limits on balance vehicles in wizard.
+* Ctrl+R to run selected code block LBM in repl.
+* Script editor search improvements.
 
 ### 6.02
 #### Released 2023-03-12